Update scripts to get live logs

+++ multiple-provers-script.sh
@@ -1,21 +1,20 @@
-# use bash shell
 #! /bin/bash
-echo "Killing existing processes on MAIN_PROVER..."
 function kill_process() {
+    # if the process is not running, don't do anything
+    if [ -z "$(ssh $2 "pgrep -f ./$1")" ]; then
+        echo "Process $1 on $2 not running!"
+        return
+    fi
     echo "Killing $1 process on $2..."
     ssh $2 "pkill -f ./$1"
     if [ $? -eq 0 ]; then
         echo "Existing process $1 on $2 killed successfully!"
     else
         echo "Existing process $1 on $2 kill failed with error code $?"
+        exit 1
     fi
 }
-# ssh MAIN_PROVER "pkill -f ./main_prover_binary"
-# if [ $? -eq 0 ]; then
-#     echo "Existing processes on MAIN_PROVER killed successfully!"
-# else
-#     echo "Existing processes on MAIN_PROVER kill failed with error code $?"
-# fi
+
 kill_process "main_prover_binary" "MAIN_PROVER"
 kill_process "sequencer_binary" "SEQUENCER"
 kill_process "prover_binary" "PROVER_1"
@@ -30,6 +29,7 @@ if [ $? -eq 0 ]; then
     echo "Sequencer binary built successfully!"
 else
     echo "Sequencer binary build failed with error code $?"
+    exit 1
 fi
 
 echo "Building main prover binary..."
@@ -38,6 +38,7 @@ if [ $? -eq 0 ]; then
     echo "Main prover binary built successfully!"
 else
     echo "Main prover binary build failed with error code $?"
+    exit 1
 fi
 
 echo "Building prover binary..."
@@ -46,17 +47,11 @@ if [ $? -eq 0 ]; then
     echo "Prover binary built successfully!"
 else
     echo "Prover binary build failed with error code $?"
+    exit 1
 fi
 
 echo "All binaries built successfully!"
 
-# transfer binaries from MAIN_PROVER to SEQUENCER, PROVER_1, PROVER_2, PROVER_3, PROVER_4
-# scp MAIN_PROVER:~/sequencer_binary SEQUENCER:~/sequencer_binary
-# if [ $? -eq 0 ]; then
-#     echo "sequencer_binary copied successfully!"
-# else
-#     echo "sequencer_binary copy failed with error code $?"
-# fi
 function copy_binary() {
     echo "Copying $1 from MAIN_PROVER to $2..."
     scp MAIN_PROVER:~/$1 $2:~/$1
@@ -64,10 +59,10 @@ function copy_binary() {
         echo "$1 copied successfully!"
     else
         echo "$1 copy failed with error code $?"
+        exit 1
     fi
 }
 copy_binary "sequencer_binary" "SEQUENCER"
-# copy_binary "main_prover_binary" "MAIN_PROVER"
 copy_binary "prover_binary" "PROVER_1"
 copy_binary "prover_binary" "PROVER_2"
 copy_binary "prover_binary" "PROVER_3"
@@ -75,8 +70,9 @@ copy_binary "prover_binary" "PROVER_4"
 
 # run binaries on MAIN_PROVER, SEQUENCER, PROVER_1, PROVER_2, PROVER_3, PROVER_4
 function run_binary() {
-    echo "Running $1 binary on $2..."
-    ssh $2 "cd ~/; nohup ./$1 > $1_log_$(date +%Y-%m-%d_%H-%M-%S).log 2>&1 & disown";
+    echo "Running $1 on $2..."
+    # $3 contains the prover id, if it is not set, then it is the sequencer or main prover
+    ssh $2 "cd ~/; nohup ./$1 $3 > $1_$2_log_$(date +%Y-%m-%d_%H-%M-%S).log 2>&1 & disown";
     if [ $? -eq 0 ]; then
         echo "$1 binary running successfully on $2!"
     else
@@ -85,11 +81,12 @@ function run_binary() {
 }
 run_binary "sequencer_binary" "SEQUENCER"
 run_binary "main_prover_binary" "MAIN_PROVER"
-run_binary "prover_binary" "PROVER_1"
-run_binary "prover_binary" "PROVER_2"
-run_binary "prover_binary" "PROVER_3"
-run_binary "prover_binary" "PROVER_4"
-# echo "Running sequencer binary..."
-# ssh SEQUENCER "cd ~/; nohup ./sequencer_binary > sequencer_log_$(date +%Y-%m-%d_%H-%M-%S).log 2>&1 & disown";
-# echo "Sequencer binary running successfully!"
+run_binary "prover_binary" "PROVER_1" "0"
+run_binary "prover_binary" "PROVER_2" "1"
+run_binary "prover_binary" "PROVER_3" "2"
+run_binary "prover_binary" "PROVER_4" "3"
 echo "All binaries running successfully!"
+
+echo "Sending generate_trace request to SEQUENCER..."
+ssh SEQUENCER "curl http://localhost:3030/block/generate_trace"
+echo "generate_trace request sent successfully!"
